Refactor: Use loader to fetch contributors data. (#791)
* Refactor to change the Router

* Fix Theme change bug

* Show Splash Screen

* Fetch Data Using loader

* Increase limit of contributors

* Remove Console.log
