+++ src/pages/index.js
@@ -1,5 +1,5 @@
 import ProjectsPage from "./ProjectsPage"
 import HomePage from "./HomePage"
-import ContriButorsPage from "./ContributorPage"
+import ContriButorsPage from "./ContributorsPage"
 import AddYourProjectsGuide from "./AddYourProjectsGuide"
 export { ProjectsPage, HomePage, ContriButorsPage, AddYourProjectsGuide }
